% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoMarvelCharacters.R
\name{InfoMarvelCharacters}
\alias{InfoMarvelCharacters}
\title{Request data of each characters from Marvel API
Including characters'name, description, comics, events and series.}
\usage{
InfoMarvelCharacters(Character, publickey, privatekey, select = "select")
}
\arguments{
\item{Character}{A character's name}

\item{publickey}{A public key you get from https://developer.marvel.com/}

\item{privatekey}{A private key you get from https://developer.marvel.com/}

\item{select}{There are three chocies of "select": "comics","series" and "events". It returns different datasets containing info of character's comics, series and events. If you want to return info of characters, you can ignore this param.}
}
\value{
A dataset. The data depends on your choice in param "select"
}
\description{
Request data of each characters from Marvel API
Including characters'name, description, comics, events and series.
}
\examples{
\dontrun{InfoMarvelCharacters("Thor",your_public_key,your_private_key)}


}
