% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoMarvelComics.R
\name{InfoMarvelComics}
\alias{InfoMarvelComics}
\title{Get Marvel Comics' Info}
\usage{
InfoMarvelComics(comics, publickey, privatekey)
}
\arguments{
\item{comics}{A comics's title}

\item{publickey}{A public key you get from https://developer.marvel.com/}

\item{privatekey}{A private key you get from https://developer.marvel.com/}
}
\value{
A dataset.
}
\description{
Request data of each comics from Marvel API
Including comics'title, description, issuenumber and format
}
\examples{
\dontrun{InfoMarvelComics("Amazing\%20Spider-Man",your_public_key,your_private_key))}

}
