% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoMarvelEvents.R
\name{InfoMarvelEvents}
\alias{InfoMarvelEvents}
\title{Get Marvel Events' Info}
\usage{
InfoMarvelEvents(events, publickey, privatekey, select = "select")
}
\arguments{
\item{events}{A events's title}

\item{publickey}{A public key you get from https://developer.marvel.com/}

\item{privatekey}{A private key you get from https://developer.marvel.com/}

\item{select}{There are three chocies of "select": "comics","series" and "events". It returns different datasets containing info of event's comics, series and characters. If you want to return info of events, you can ignore this param.}
}
\value{
A dataset. The data depends on your choice in param "select"
}
\description{
Request data of each events from Marvel API
Including event's title, description, start year and end year
A user can select more info of comics, characters and series related to the events as well.
}
\examples{
\dontrun{InfoMarvelEvents("Civil\%20War",your_public_key,your_private_key)}

}
