% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoMarvelSeries.R
\name{InfoMarvelSeries}
\alias{InfoMarvelSeries}
\title{Get Marvel Series' Info}
\usage{
InfoMarvelSeries(series, publickey, privatekey)
}
\arguments{
\item{series}{A series's title}

\item{publickey}{A public key you get from https://developer.marvel.com/}

\item{privatekey}{A private key you get from https://developer.marvel.com/}
}
\value{
A dataset.
}
\description{
Request data of each series from Marvel API
Including series'title, description, start year, end year, rating and type
}
\examples{
\dontrun{InfoMarvelSeries("Black\%20Panther",your_public_key,your_private_key)}

}
