% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarvelCharacters.R
\name{MarvelCharacters}
\alias{MarvelCharacters}
\title{Get ALL Marvel Characters' Info}
\usage{
MarvelCharacters(publickey, privatekey)
}
\arguments{
\item{publickey}{A public key you get from https://developer.marvel.com/}

\item{privatekey}{A private key you get from https://developer.marvel.com/}
}
\value{
A dataset including characters' name and short description.
}
\description{
Request data from Marvel API
to get a dataset including ALL Marvel characters'name and short description (total 1491)
}
\examples{
\dontrun{MarvelCharacters(your_public_key,your_private_key)}

}
\keyword{Marvel,Characters,data}
